refactor(knowledge-artifacts): 收回通用来源解析并删除重复实现
