+++ src/knowledge-artifacts/governance/source-probe.ts
@@ -1,6 +1,6 @@
 import { existsSync, lstatSync, readdirSync, type Dirent } from 'node:fs';
 import { isAbsolute, resolve, join } from 'node:path';
-import { resolveInstallSource } from '../../eval-workflows/inputs/source-resolver.js';
+import { resolveInstallSource } from '../sources/install-source.js';
 import { hashArtifactSource, isDistributablePath } from '../sources/content-hash.js';
 import type { SourceProbe } from './list-view.js';
 import type { ManagedArtifactRecord } from './contracts.js';
